Add day18, which seems to have some problem with part 2
I've been over it pretty carefully, and can't see the issue. There's clearly
a repeating pattern starting at T=501, with period=28, and
(1e9 - 501) % 28 = 23, which means score[501+23] should be the same as
score[1e9]. But it is rejected, and so are the scores before and after so
its not an off-by-one...so I'll have to come back later.